+++ frontend/src/components/message-input/index.ts
@@ -1,0 +1,6 @@
+export { default as AttachFileButton } from './AttachFileButton';
+export { default as AttachmentChipList } from './AttachmentChipList';
+export { default as MicButton } from './MicButton';
+export { default as SourcesTrigger } from './SourcesTrigger';
+export { default as ToolsTrigger } from './ToolsTrigger';
+export type { RecordingState } from './MicButton';
